Fix: user 페이지 관련 오류 및 버그 수정
