+++ src/data/siteCard.json
@@ -1,0 +1,102 @@
+[
+  {
+    "id": 1,
+    "title": "Card 1231231231231232131231231231232132",
+    "description": "숙박 예약 사이트 1132123312213132132132",
+    "image": "/public/images/possible-1.png",
+    "link": "https://example.com/card1",
+    "rating": 4.5,
+    "ratingParticipation": 100,
+    "viewer": 200
+  },
+  {
+    "id": 2,
+    "title": "Card 1",
+    "description": "숙박 예약 사이트 1",
+    "image": "/public/images/possible-1.png",
+    "link": "https://example.com/card1",
+    "rating": 4.5,
+    "ratingParticipation": 100,
+    "viewer": 200
+  },
+  {
+    "id": 3,
+    "title": "Card 1",
+    "description": "숙박 예약 사이트 1",
+    "image": "/public/images/possible-1.png",
+    "link": "https://example.com/card1",
+    "rating": 4.5,
+    "ratingParticipation": 100,
+    "viewer": 200
+  },
+  {
+    "id": 4,
+    "title": "Card 1",
+    "description": "숙박 예약 사이트 1",
+    "image": "/public/images/possible-1.png",
+    "link": "https://example.com/card1",
+    "rating": 4.5,
+    "ratingParticipation": 100,
+    "viewer": 200
+  },
+  {
+    "id": 5,
+    "title": "Card 1",
+    "description": "숙박 예약 사이트 1",
+    "image": "/public/images/possible-1.png",
+    "link": "https://example.com/card1",
+    "rating": 4.5,
+    "ratingParticipation": 100,
+    "viewer": 200
+  },
+  {
+    "id": 6,
+    "title": "Card 1",
+    "description": "숙박 예약 사이트 1",
+    "image": "/public/images/possible-1.png",
+    "link": "https://example.com/card1",
+    "rating": 4.5,
+    "ratingParticipation": 100,
+    "viewer": 200
+  },
+  {
+    "id": 7,
+    "title": "Card 1",
+    "description": "숙박 예약 사이트 1",
+    "image": "/public/images/possible-1.png",
+    "link": "https://example.com/card1",
+    "rating": 4.5,
+    "ratingParticipation": 100,
+    "viewer": 200
+  },
+  {
+    "id": 8,
+    "title": "Card 1",
+    "description": "숙박 예약 사이트 1",
+    "image": "/public/images/possible-1.png",
+    "link": "https://example.com/card1",
+    "rating": 4.5,
+    "ratingParticipation": 100,
+    "viewer": 200
+  },
+  {
+    "id": 9,
+    "title": "Card 1",
+    "description": "숙박 예약 사이트 1",
+    "image": "/public/images/possible-1.png",
+    "link": "https://example.com/card1",
+    "rating": 4.5,
+    "ratingParticipation": 100,
+    "viewer": 200
+  },
+  {
+    "id": 10,
+    "title": "Card 1",
+    "description": "숙박 예약 사이트 1",
+    "image": "/public/images/possible-1.png",
+    "link": "https://example.com/card1",
+    "rating": 4.5,
+    "ratingParticipation": 100,
+    "viewer": 200
+  }
+]
